+++ ui/desktop/src/utils/pathUtils.ts
@@ -1,67 +1,5 @@
 import path from 'node:path';
-import fs from 'node:fs';
 import os from 'node:os';
-import Electron from 'electron';
-import log from './logger';
-
-export const getGoosedBinaryPath = (app: Electron.App): string => {
-  let executableName = process.platform === 'win32' ? 'goosed.exe' : 'goosed';
-
-  const possiblePaths: string[] = [];
-  addPaths(false, possiblePaths, executableName, app);
-
-  for (const binPath of possiblePaths) {
-    try {
-      const resolvedPath = path.resolve(binPath);
-
-      if (fs.existsSync(resolvedPath)) {
-        const stats = fs.statSync(resolvedPath);
-        if (stats.isFile()) {
-          return resolvedPath;
-        } else {
-          log.error(`Path exists but is not a regular file: ${resolvedPath}`);
-        }
-      }
-    } catch (error) {
-      log.error(`Error checking path ${binPath}:`, error);
-    }
-  }
-
-  throw new Error(
-    `Could not find ${executableName} binary in any of the expected locations: ${possiblePaths.join(
-      ', '
-    )}`
-  );
-};
-
-const addPaths = (
-  isWindows: boolean,
-  possiblePaths: string[],
-  executableName: string,
-  app: Electron.App
-): void => {
-  if (!app.isPackaged) {
-    possiblePaths.push(
-      path.join(process.cwd(), 'src', 'bin', executableName),
-      path.join(process.cwd(), 'bin', executableName),
-      path.join(process.cwd(), '..', '..', 'target', 'debug', executableName),
-      path.join(process.cwd(), '..', '..', 'target', 'release', executableName)
-    );
-  } else {
-    possiblePaths.push(
-      path.join(process.resourcesPath, 'bin', executableName),
-      path.join(app.getAppPath(), 'resources', 'bin', executableName)
-    );
-
-    if (isWindows) {
-      possiblePaths.push(
-        path.join(process.resourcesPath, executableName),
-        path.join(app.getAppPath(), 'resources', executableName),
-        path.join(app.getPath('exe'), '..', 'bin', executableName)
-      );
-    }
-  }
-};
 
 /**
  * Expands tilde (~) to the user's home directory
